done with trainerComments

+++ src/routers/trainerComments.js
@@ -1,7 +1,7 @@
 const Joi = require('joi')
 const express = require('express');
 const dbOperations = require('../database/operations');
-const { authenticateUser } = require('../middleware/authentication');
+const { authenticateUser, ADMIN_ROLE } = require('../middleware/authentication');
 const { validateId } = require('../middleware/common');
 const { validateTrainerComment, validateCommentExists } = require('../middleware/trainerComments');
 const { validateTrainerExists } = require('../middleware/trainers');
@@ -56,12 +56,18 @@ router.put('/api/trainers/comments/:id',
         validateTrainerComment,
         validateCommentExists
     ], async (req, res) => {
+    const existingComment = req.commentObject;
+    const user = req.user
     const commentId = req.params.id
-    const comment = req.comment
+    const newComment = req.comment
     
     try {
-        await dbOperations.updateComment(comment, commentId)
-        res.sendStatus(200)
+        if (existingComment.fk_user_id == user.id || user.role == ADMIN_ROLE) {
+            await dbOperations.updateComment(newComment, commentId)
+            res.sendStatus(200)
+        } else {
+            res.sendStatus(403)
+        }        
     } catch (e) {
         console.log(e)
         res.sendStatus(500)
@@ -74,10 +80,16 @@ router.delete('/api/trainers/comments/:id',
         validateId,
         validateCommentExists
     ], async (req, res) => {
+    const existingComment = req.commentObject;
+    const user = req.user
     const commentId = req.params.id
     try {
-        await dbOperations.deleteComment(commentId)
-        res.sendStatus(204)
+        if (existingComment.fk_user_id == user.id || user.role == ADMIN_ROLE) {
+            await dbOperations.deleteComment(commentId)
+            res.sendStatus(204)
+        } else {
+            res.sendStatus(403)
+        }
     } catch (e) {
         console.log(e)
         res.sendStatus(500)
